requirements.txt
requests
python-whois
builtwith
bs4
os
typer